Access to Insurance Data

This function returns a data frame about insurance demographics from State Cancer Profiles.

demo_insurance(area, areatype, insurance, sex, age, race = NULL)

Arguments

area

A state/territory abbreviation or USA.

areatype

One of the following values:

"county"

"hsa" (Health Service Area)

"state".

insurance

One of the following values:

"% Insured in demographic group, all income levels"

"% Insured in demographic group, people at or below 138% of Poverty"

"% Insured in demographic group, people at or below 200% of Poverty"

"% Insured in demographic group, people at or below 250% of Poverty"

"% Insured in demographic group, people at or below 400% of Poverty"

"% Insured in demographic group, people between 138% - 400% of poverty"

"% uninsured in demographic group, all income levels"

"% uninsured in demographic group, people at or below 138% of Poverty"

"% uninsured in demographic group, people at or below 200% of Poverty"

"% uninsured in demographic group, people at or below 250% of Poverty"

"% uninsured in demographic group, people at or below 400% of Poverty"

"% uninsured in demographic group, people between 138% - 400% of poverty".

sex

One of the following values:

"both sexes"

"male"

"female".

age

If you specified "both sexes" for sex choose one of the following values:

"under 19 years"

"18 to 64 years"

"21 to 64 years"

"40 to 64 years"

"50 to 64 years"

"under 65 years".

Otherwise if you specified "male" or "female" for sex, choose one of the following values:

"18 to 64 years"

"40 to 64 years"

"50 to 64 years"

"Under 65 years".

race

Only specify race if you specified "state" for areatype

"All Races (includes Hispanic)"

"White (non-Hispanic)"

"black (non-Hispanic)"

"American Indian / Alaska Native (non-Hispanic)"

"Asian (non-Hispanic)"

"Hispanic (Any Race)".

Value

A data frame with the following columns: Area Type, Area Code, Percent, People, Rank.

See also

Examples

if (FALSE) {
demo_insurance(
  area = "usa",
  areatype = "state",
  insurance = "% Insured in demographic group, all income levels",
  sex = "both sexes",
  age = "18 to 64 years",
  race = "white (non-hispanic)"
)

demo_insurance(
  area = "wa",
  areatype = "hsa",
  insurance = "% Insured in demographic group, all income levels",
  sex = "males",
  age = "18 to 64 years"
)

demo_insurance(
  area = "dc",
  areatype = "county",
  insurance = "% Insured in demographic group, all income levels",
  sex = "males",
  age = "18 to 64 years"
)
}