Access to Cancer Mortality Data

This function returns a data frame about cancer mortality from State Cancer Profiles.

mortality_cancer(area, areatype, cancer, race, sex, age, year)

Arguments

area

A state/territory abbreviation or USA.

areatype

One of the following values:

"county"

"hsa" (Health Service Area)

"state".

cancer

One of the following values:

"all cancer sites"

"bladder"

"brain & ons"

"breast (female)"

"cervix"

"childhood (ages <15, all sites)"

"childhood (ages <20, all sites)"

"colon & rectum"

"esophagus"

"kidney & renal pelvis"

"leukemia"

"liver & bile duct"

"lung & bronchus"

"melanoma of the skin"

"non-hodgkin lymphoma"

"oral cavity & pharynx"

"ovary"

"pancreas"

"prostate"

"stomach"

"thyroid"

"uterus (corpus & uterus, nos)"

race

One of the following values:

"All Races (includes Hispanic)"

"White (non-Hispanic)"

"Black (non-Hispanic)"

"American Indian / Alaska Native (non-Hispanic)"

"Asian / Pacific Islander (non-Hispanic)"

"Hispanic (Any Race)".

sex

One of the following values:

"both sexes"

"male"

"female".

age

One of the following values:

"all ages"

"ages <50"

"ages 50+"

"ages <65"

"ages 65+"

"ages <15"

"ages <20".

year

One of the following values:

"latest 5 year average"

"latest single year (us by state)".

Value

A data frame with the following columns: Area Type, Area Code, Met Healthy People Objective of ***?, Age Adjusted Death Rate, Lower 95% CI Rate, Upper 95% CI Rate, CI Rank, Lower CI Rank, Upper CI Rank, Annual Average Count, Recent Trend, Recent 5 Year Trend, Lower 95% CI Trend, Upper 95% CI Trend.

Examples

if (FALSE) {
mortality_cancer(
  area = "wa",
  areatype = "county",
  cancer = "all cancer sites",
  race = "black (non-hispanic)",
  sex = "both sexes",
  age = "ages 65+",
  year = "latest 5 year average"
)

mortality_cancer(
  area = "usa",
  areatype = "state",
  cancer = "prostate",
  race = "all races (includes hispanic)",
  sex = "males",
  age = "ages 50+",
  year = "latest single year (us by state)"
)

mortality_cancer(
  area = "wa",
  areatype = "hsa",
  cancer = "ovary",
  race = "all races (includes hispanic)",
  sex = "females",
  age = "ages 50+",
  year = "latest 5 year average"
)
}