Access to Smoking Data

This function returns a data frame about smoking risks from State Cancer Profiles.

Usage

risk_smoking(smoking, race = NULL, sex = NULL, datatype = NULL, area = NULL)

Arguments

smoking

The only permissible values are

"smoking laws (any)"

"smoking laws (bars)"

"smoking laws (restaurants)"

"smoking laws (workplace)"

"smoking laws (workplace; restaurant; & bar)"

"smokers (stopped for 1 day or longer)"

"smoking not allowed at work (all people)"

"smoking not allowed in home (all people)"

"smoking not allowed at work (current smokers)"

"smoking not allowed at work (former/never smokers)"

"smoking not allowed in home (current smokers)"

"smoking not allowed in home (former/never smokers)"

"former smoker; ages 18+"

"former smoker, quit 1 year+; ages 18+"

"smokers (ever); ages 18+"

"e-cigarette use; ages 18+"

"smokers (current); ages 18+".

race

One of the following values:

"All Races (includes Hispanic)"

"White (non-Hispanic)"

"Black (non-Hispanic)"

"American Indian / Alaska Native (non-Hispanic)"

"Asian / Pacific Islander (non-Hispanic)"

"Hispanic (Any Race)".

sex

One of the following values:

"both sexes"

"male"

"female".

datatype

One of the following values:

"direct estimates"

"county level modeled estimates".

area

A state/territory abbreviation or USA.

Value

A data frame with the following columns: Area Type, Area Code, Percent, Lower CI 95%, Upper CI 95%, Number of Respondents.

Details

Please note that this function requires very specific arguments for each smoking type.

See also

Examples

if (FALSE) {
risk_smoking(smoking = "smoking laws (any)")

risk_smoking(
  smoking = "smokers (stopped for 1 day or longer)",
  sex = "both sexes",
  datatype = "county level modeled estimates",
  area = "wa"
)

risk_smoking(
  smoking = "smoking not allowed at work (current smokers)",
  sex = "both sexes",
  datatype = "direct estimates"
)

risk_smoking(
  smoking = "smokers (current); ages 18+",
  race = "all races (includes hispanic)",
  sex = "both sexes",
  datatype = "county level modeled estimates",
  area = "wa"
)
}